Moved the CubeService to CubeService.java so that a description could be set in the Service that will appear in the CDAP UI. Fixed a broken query. Corrected copyright dates. Further edits to README.
