+++ resources/query-ip-browser.json
@@ -5,7 +5,7 @@
     "endTs":   1423398198,
     "measureNames": ["bytes.sent"],
     "measureType": "COUNTER",
-    "sliceByTagValues": [{"name": "ip", "value": "69.181.160.120"}],
+    "sliceByTagValues": {"ip": "69.181.160.120"},
     "groupByTags": ["browser"],
     "limit": 1000
 }
